+++ src/views/Songs.vue
@@ -64,7 +64,7 @@ const options = {
 };
 
 axios.request(options).then(function (response) {
-	console.log(response.data);
+	// console.log(response.data);
 }).catch(function (error) {
 	console.error(error);
 });
